ISUE #1691: Pace data smoothing for activity charts
