+++ Makefile
@@ -56,14 +56,11 @@ translation-debug:
 phpunit:
 	@make dcr cmd="vendor/bin/phpunit --order-by=random $(arg)"
 
-phpunit-with-coverage-report:
-	@make phpunit arg="--coverage-clover=clover.xml -d --min-coverage=min-coverage-rules.php -d --clean-up-clover-xml"
-
-phpunit-html-coverage:
-	@make phpunit arg="--coverage-html var/coverage"
-
 paratest:
 	@make dcr cmd="vendor/bin/paratest --order-by=random --processes=auto $(arg)"
+
+phpunit-html-coverage:
+	@make paratest arg="--coverage-html var/coverage"
 
 phpstan:
 	@make dcr cmd="vendor/bin/phpstan --memory-limit=1G $(arg)"
