+++ public/css/tailwind.output.css
@@ -1078,9 +1078,6 @@
   .mx-auto {
     margin-inline: auto;
   }
-  .my-4 {
-    margin-block: calc(var(--spacing) * 4);
-  }
   .ms-2 {
     margin-inline-start: calc(var(--spacing) * 2);
   }
@@ -4704,6 +4701,9 @@
           @media (hover: hover) {
             color: var(--color-strava-orange);
           }
+        }
+        &.active {
+          color: var(--color-strava-orange);
         }
       }
     }
