+++ tests/Domain/Activity/MathTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Tests\Domain\Dashboard\Widget\AthleteProfile;
+namespace App\Tests\Domain\Activity;
 
-use App\Domain\Dashboard\Widget\AthleteProfile\Math;
+use App\Domain\Activity\Math;
 use PHPUnit\Framework\Attributes\TestWith;
 use PHPUnit\Framework\TestCase;
 
@@ -18,10 +18,20 @@ class MathTest extends TestCase
         );
     }
 
-    public function testItShouldThrowWhenValuesAreEmpty(): void
+    public function testMedianItShouldThrowWhenValuesAreEmpty(): void
     {
         $this->expectExceptionObject(new \InvalidArgumentException('Cannot calculate median of empty array.'));
 
         Math::median([]);
+    }
+
+    #[TestWith(data: [[], []])]
+    #[TestWith(data: [[1, 2, 3, 4], [2, 2.5, 2.5, 3]])]
+    public function testMovingAverage(array $values, array $expectedResult): void
+    {
+        $this->assertEquals(
+            $expectedResult,
+            Math::movingAverage($values, 5)
+        );
     }
 }
